requirements.txt
flask
redis<3.0.0
google-cloud-storage


